Changed wttr to an alias
It used to used to be a command that would run on reloading the window
manager. This means that the command needed to be displayed so it was
done in a terminal, urxvt. When the command was executed through a
terminal, not all of the http data transmits and cuts off after the
terminal believes that it is done executing. Nothing keeps the process
runnig so it dies. The other options are ignored. Also, it is a tad
annoying to see the weather every time I want to refresh the config.